# ApiAulearning/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aulearning API",
        "description": "API de Aulearning",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/enrollments": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Listar inscripciones",
                "operationId": "46f86247fae90db42c50de2b294234ad",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "course_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "enrollment_date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "enrollment_date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortDirection"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de inscripciones"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Crear inscripción",
                "operationId": "8b49ef3dceb0238cfbd9996b169728e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "course_id"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscripción creada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/enrollments/{id}": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Ver inscripción",
                "operationId": "7702a76cd3c43405c6aae82e79572999",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscripción encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Eliminar inscripción",
                "operationId": "7f6eede96ccbc31ddf0f9cbb9410466a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscripción eliminada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Número de elementos por página",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 15,
                    "minimum": 1
                }
            },
            "SortBy": {
                "name": "sort_by",
                "in": "query",
                "description": "Campo por el que ordenar",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "id"
                }
            },
            "SortDirection": {
                "name": "sort_direction",
                "in": "query",
                "description": "Dirección de ordenación",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "desc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}